pyproject.toml
[project]
name = "upwork-proposals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "langchain-openai>=1.1.16",
    "playwright>=1.58.0",
    "pydantic-settings>=2.14.0",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "twilio>=9.10.5",
    "uvicorn>=0.45.0",
]
